move user.css to styles folder
