+++ client/src/components/users/Register.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 import { register } from './userActions';
-import "./user.css";
+import "../../styles/user.css";
 
 function Register(props) {
