+++ server/server.js
@@ -56,5 +56,5 @@ app.use(function(err, req, res, next) {
 // Starting Server
 const PORT = config.PORT;
 app.listen(PORT, () => {
-  console.log(`🌎 ==> Server now listening on PORT ${PORT}!`);
+  console.log(`🌎 ==> Server listening on PORT ${PORT}!`);
 });
